+++ src/headers/aoc11.h
@@ -1,7 +1,7 @@
 #ifndef __AOC11_H_
 #define __AOC11_H_
 
-unsigned int aoc11( const char* filepath );
-unsigned int aoc11_2( const char* filepath );
+unsigned long long int aoc11( const char* filepath );
+unsigned long long int aoc11_2( const char* filepath );
 
 #endif
