+++ src/headers/aoc12.h
@@ -4,7 +4,7 @@
 #ifndef __AOC12_H_
 #define __AOC12_H_
 
-int aoc12( const char* filepath );
-int aoc12_2( const char* filepath );
+unsigned long long int aoc12( const char* filepath );
+unsigned long long int aoc12_2( const char* filepath );
 
 #endif
