+++ src/headers/aoc14.h
@@ -4,7 +4,7 @@
 #ifndef __AOC14_H_
 #define __AOC14_H_
 
-int aoc14( const char* filepath );
-int aoc14_2( const char* filepath );
+unsigned long long int aoc14( const char* filepath );
+unsigned long long int aoc14_2( const char* filepath );
 
 #endif
