+++ src/headers/aoc3.h
@@ -1,7 +1,7 @@
 #ifndef __AOC3_H_
 #define __AOC3_H_
 
-int aoc3( const char* filepath );
-int aoc3_2( const char* filepath );
+unsigned long long int aoc3( const char* filepath );
+unsigned long long int aoc3_2( const char* filepath );
 
 #endif
