+++ src/headers/aoc7.h
@@ -1,7 +1,7 @@
 #ifndef __AOC7_H_
 #define __AOC7_H_
 
-int aoc7( const char* filepath );
-int aoc7_2( const char* filepath );
+unsigned long long int aoc7( const char* filepath );
+unsigned long long int aoc7_2( const char* filepath );
 
 #endif
