+++ src/headers/aoc8.h
@@ -1,7 +1,7 @@
 #ifndef __AOC8_H_
 #define __AOC8_H_
 
-int aoc8( const char* filepath );
-int aoc8_2( const char* filepath );
+unsigned long long int aoc8( const char* filepath );
+unsigned long long int aoc8_2( const char* filepath );
 
 #endif
