+++ src/headers/aoc9.h
@@ -1,7 +1,7 @@
 #ifndef __AOC9_H_
 #define __AOC9_H_
 
-int aoc9( const char* filepath );
-int aoc9_2( const char* filepath );
+unsigned long long int aoc9( const char* filepath );
+unsigned long long int aoc9_2( const char* filepath );
 
 #endif
